Remove deprecated quality tools
