--- src/main/java/org/magicdgs/readtools/tools/quality/StandardizeQuality.java
@@ -1,160 +1,0 @@
-/*
- * The MIT License (MIT)
- *
- * Copyright (c) 2015 Daniel Gómez-Sánchez
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
-package org.magicdgs.readtools.tools.quality;
-
-import org.magicdgs.io.readers.bam.SamReaderSanger;
-import org.magicdgs.io.readers.fastq.single.FastqReaderSingleInterface;
-import org.magicdgs.io.readers.fastq.single.FastqReaderSingleSanger;
-import org.magicdgs.readtools.cmd.ReadToolsLegacyArgumentDefinitions;
-import org.magicdgs.readtools.cmd.programgroups.DeprecatedProgramGroup;
-import org.magicdgs.readtools.tools.ReadToolsBaseTool;
-import org.magicdgs.readtools.utils.fastq.QualityUtils;
-import org.magicdgs.readtools.utils.logging.FastqLogger;
-import org.magicdgs.readtools.utils.logging.ProgressLoggerExtension;
-import org.magicdgs.readtools.utils.misc.IOUtils;
-import org.magicdgs.readtools.utils.read.ReadReaderFactory;
-import org.magicdgs.readtools.utils.read.ReadWriterFactory;
-
-import htsjdk.samtools.SAMFileHeader;
-import htsjdk.samtools.SAMFileWriter;
-import htsjdk.samtools.SAMRecord;
-import htsjdk.samtools.SamReader;
-import htsjdk.samtools.fastq.FastqRecord;
-import htsjdk.samtools.fastq.FastqWriter;
-import htsjdk.samtools.util.CloserUtil;
-import htsjdk.samtools.util.FastqQualityFormat;
-import org.broadinstitute.barclay.argparser.Argument;
-import org.broadinstitute.barclay.argparser.CommandLineException;
-import org.broadinstitute.barclay.argparser.CommandLineProgramProperties;
-import org.broadinstitute.hellbender.exceptions.UserException;
-
-import java.io.Closeable;
-import java.io.File;
-
-/**
- * Class for converting from Illumina to Sanger encoding both FASTQ and BAM files
- *
- * @author Daniel Gomez-Sanchez (magicDGS)
- * @deprecated this tool correspond to legacy elements.
- */
-@CommandLineProgramProperties(oneLineSummary = "DEPRECATED: USE 'StandardizeReads' for convert an Illumina BAM/FASTQ into Sanger.",
-        summary = "DEPRECATED: USE 'StandardizeReads' instead.\n" +
-                "The standard encoding for a BAM file is Sanger and this tool is provided to standardize both BAM/FASTQ files "
-                        + "for latter analysis. It does not support mixed qualities.",
-        programGroup = DeprecatedProgramGroup.class)
-@Deprecated
-public final class StandardizeQuality extends ReadToolsBaseTool {
-
-    @Argument(fullName = ReadToolsLegacyArgumentDefinitions.INPUT_LONG_NAME, shortName = ReadToolsLegacyArgumentDefinitions.INPUT_SHORT_NAME, optional = false,
-            doc = "Input BAM/FASTQ to determine the quality.")
-    public File input = null;
-
-    @Argument(fullName = ReadToolsLegacyArgumentDefinitions.OUTPUT_LONG_NAME, shortName = ReadToolsLegacyArgumentDefinitions.OUTPUT_SHORT_NAME, optional = false,
-            doc = "Output for the converted file. The extension determine the format SAM/BAM or FASTQ/GZIP.")
-    public String output = null;
-
-    private Closeable reader;
-    private Closeable writer;
-    private ReadWriterFactory factory;
-
-    @Override
-    protected String[] customCommandLineValidation() {
-        if (maintainFormat) {
-            throw new CommandLineException.BadArgumentValue(
-                    ReadToolsLegacyArgumentDefinitions.MAINTAIN_FORMAT_SHORT_NAME);
-        }
-        if (QualityUtils.getFastqQualityFormat(input) == FastqQualityFormat.Standard) {
-            throw new UserException.BadInput(
-                    String.format("%s already in Sanger formatting", input));
-        }
-        return super.customCommandLineValidation();
-    }
-
-    @Override
-    protected Object doWork() {
-        factory = new ReadWriterFactory()
-                .setReferenceFile(REFERENCE_SEQUENCE)
-                .setCreateIndex(CREATE_INDEX)
-                .setCreateMd5File(CREATE_MD5_FILE)
-                .setUseAsyncIo(nThreads != 1)
-                .setForceOverwrite(forceOverwrite);
-        if (IOUtils.isSamBamOrCram(input.toPath())) {
-            runBam();
-        } else {
-            if (CREATE_INDEX) {
-                logger.warn("Index won't be performed for FASTQ output file.");
-            }
-            runFastq();
-        }
-        return null;
-    }
-
-
-    /**
-     * Change the format in a Fastq file
-     */
-    private void runFastq() {
-        // open reader (directly converting)
-        final FastqReaderSingleInterface reader =
-                new FastqReaderSingleSanger(input, allowHigherSangerQualities);
-        // open writer
-        final FastqWriter writer = factory.openFastqWriter(output);
-        // start iterations
-        final FastqLogger progress = new FastqLogger(logger);
-        for (final FastqRecord record : reader) {
-            writer.write(record);
-            progress.add();
-        }
-        progress.logNumberOfVariantsProcessed();
-        this.reader = reader;
-        this.writer = writer;
-    }
-
-    /**
-     * Change the format in a BAM file
-     */
-    private void runBam() {
-        final SamReader reader =
-                new SamReaderSanger(input, new ReadReaderFactory(), allowHigherSangerQualities);
-        final SAMFileHeader header = reader.getFileHeader();
-        header.addProgramRecord(getToolProgramRecord());
-        final SAMFileWriter writer = factory.openSAMWriter(header, true, output);
-        // start iterations
-        final ProgressLoggerExtension progress = new ProgressLoggerExtension(logger);
-        for (final SAMRecord record : reader) {
-            writer.addAlignment(record);
-            progress.record(record);
-        }
-        progress.logNumberOfVariantsProcessed();
-        this.reader = reader;
-        this.writer = writer;
-    }
-
-    @Override
-    protected void onShutdown() {
-        CloserUtil.close(writer);
-        CloserUtil.close(reader);
-    }
-}
